code for user login as well as testing
Checks if a user is logged in by verifying the presence of the user_id session variable.
Retrieves user data from the database based on the user_id.
If user data is found, it returns the data; otherwise, it redirects to the login page.
Calls the check_login() function to determine if the user is logged in.
If logged in, user data is returned; otherwise, the user is redirected to the login page.
Validate the provided credentials against the database.
If valid, authenticate the user and start their session.
Redirect them to the appropriate page.
If invalid, display an error message prompting the user to try again or register.